incremental refactoring for client lib
