+++ broke-sample/src/main/java/org/dsngroup/broke/MultiClient.java
@@ -17,7 +17,6 @@
 package org.dsngroup.broke;
 
 import org.dsngroup.broke.client.BlockClient;
-import org.dsngroup.broke.client.ConnectDeniedException;
 import org.dsngroup.broke.protocol.MqttQoS;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -57,84 +56,82 @@ public class MultiClient {
             System.exit(1);
         }
     }
-}
 
-class PublishClient extends Thread {
+    private static class PublishClient extends Thread {
 
-    private BlockClient blockClient;
+        private BlockClient blockClient;
 
-    private static final Logger logger = LoggerFactory.getLogger(PublishClient.class);
+        private static final Logger logger = LoggerFactory.getLogger(PublishClient.class);
 
-    private String topic;
+        private String topic;
 
-    private int sleepInterval;
+        private int sleepInterval;
 
-    @Override
-    public void run() {
-        try {
-            blockClient.connect(1, 0);
-            while(true) {
-                blockClient.publish(topic, MqttQoS.AT_LEAST_ONCE, 0, "Foo");
-                Thread.sleep(sleepInterval);
+        @Override
+        public void run() {
+            try {
+                blockClient.connect(1, 0);
+                while (true) {
+                    blockClient.publish(topic, MqttQoS.AT_LEAST_ONCE, 0, "Foo");
+                    Thread.sleep(sleepInterval);
+                }
+            } catch (Exception e) {
+                logger.error(e.getMessage());
+                return;
             }
-        } catch(Exception e) {
-            logger.error(e.getMessage());
-            logger.error(e.getStackTrace().toString());
-            return;
+
         }
 
+        public PublishClient(String serverAddress, int serverPort, String topic, int sleepInterval) {
+            try {
+                this.blockClient = new BlockClient(serverAddress, serverPort);
+                this.topic = topic;
+                this.sleepInterval = sleepInterval;
+            } catch (Exception e) {
+                logger.error(e.getMessage());
+                return;
+            }
+
+        }
     }
 
-    public PublishClient(String serverAddress, int serverPort, String topic, int sleepInterval) {
-        try {
-            this.blockClient = new BlockClient(serverAddress, serverPort);
-            this.topic = topic;
-            this.sleepInterval = sleepInterval;
-        } catch (Exception e){
-            logger.error(e.getMessage());
-            logger.error(e.getStackTrace().toString());
-            return;
+    private static class SubscribeClient extends Thread {
+
+        private BlockClient blockClient;
+
+        private static final Logger logger = LoggerFactory.getLogger(SubscribeClient.class);
+
+        private String topicFilter;
+
+        private int groupId;
+
+        @Override
+        public void run() {
+            try {
+                blockClient.connect(1, 0);
+                blockClient.subscribe(topicFilter, MqttQoS.AT_LEAST_ONCE, 0, groupId);
+            } catch (Exception e) {
+                logger.error(e.getMessage());
+                return;
+            }
+
         }
 
-    }
-}
+        public SubscribeClient(String serverAddress, int serverPort, String topicFilter, int groupId) {
 
-class SubscribeClient extends Thread {
+            try {
+                this.blockClient = new BlockClient(serverAddress, serverPort);
+                this.topicFilter = topicFilter;
+                this.groupId = groupId;
+            } catch (ArrayIndexOutOfBoundsException e) {
+                logger.error("Not enough arguments");
+                System.exit(1);
+            } catch (Exception e) {
+                logger.error(e.getMessage());
+                return;
+            }
 
-    private BlockClient blockClient;
-
-    private static final Logger logger = LoggerFactory.getLogger(SubscribeClient.class);
-
-    private String topicFilter;
-
-    private int groupId;
-
-    @Override
-    public void run() {
-        try {
-            blockClient.connect(1, 0);
-            blockClient.subscribe(topicFilter, MqttQoS.AT_LEAST_ONCE, 0, groupId);
-        } catch (Exception e) {
-            logger.error(e.getMessage());
-            logger.error(e.getStackTrace().toString());
-            return;
         }
-
     }
 
-    public SubscribeClient(String serverAddress, int serverPort, String topicFilter, int groupId) {
-
-        try {
-            this.blockClient = new BlockClient(serverAddress, serverPort);
-            this.topicFilter = topicFilter;
-            this.groupId = groupId;
-        } catch (ArrayIndexOutOfBoundsException e) {
-            logger.error("Not enough arguments");
-            System.exit(1);
-        } catch (Exception e) {
-            logger.error(e.getMessage());
-            return;
-        }
-
-    }
 }
